pyproject.toml
[tool.poetry]
name = "ditti-cms-commercial"
version = "0.1.0"
description = ""
authors = ["Alex Paden <padenfool@gmail.com>"]
readme = "README.md"
packages = [{include = "ditti_web"}]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
farcaster = "^0.7.0"
python-dotenv = "^0.21.1"
supabase = "^1.0.0"
flask-migrate = "^4.0.4"
psycopg2-binary = "^2.9.5"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
